# ⚡️ Speed up function `should_resolve_list_connection_edges` by 46%