@patrick91 2025-08-19
@bellini666 can you check this? 

I also wanted to get some benchmarks to understand when this optimisation is actually effective, from a quick test with claude it might not be faster in _all_ cases 🤔