@bellini666 2025-08-20
@patrick91 this time I like it :)

The code actually looks better to read as well, if not, I would probably say that the speed improvement would again be negligible.